+++ .gitignore
@@ -22,6 +22,7 @@ app/chip_debug/release/
 # testing
 /*.conf
 /*.yaml
+!/.pre-commit-config.yaml
 /test
 /build
