+++ web/apps/customerportal/src/app/routes/topup/TopUp.tsx
@@ -28,26 +28,30 @@ declare global {
 
 type TopUpState =
   | { stage: "initial" }
-  | { stage: "sumup"; topupAmount: number; checkoutId: string }
+  | { stage: "sumup"; topupAmount: number; checkoutId: string; orderUUID: string }
   | { stage: "success" }
   | { stage: "error"; message?: string };
 
 const initialState: TopUpState = { stage: "initial" };
 
 type TopUpStateAction =
-  | { type: "created-checkout"; topupAmount: number; checkoutId: string }
+  | { type: "created-checkout"; topupAmount: number; checkoutId: string; orderUUID: string }
   | { type: "sumup-success" }
   | { type: "sumup-error"; message?: string }
   | { type: "reset" };
 
 const reducer = (state: TopUpState, action: TopUpStateAction): TopUpState => {
-  console.log("processing action", action, "prev state", state);
   switch (action.type) {
     case "created-checkout":
       if (state.stage !== "initial") {
         return state;
       }
-      return { stage: "sumup", topupAmount: action.topupAmount, checkoutId: action.checkoutId };
+      return {
+        stage: "sumup",
+        topupAmount: action.topupAmount,
+        checkoutId: action.checkoutId,
+        orderUUID: action.orderUUID,
+      };
     case "sumup-success":
       if (state.stage !== "sumup") {
         return state;
@@ -97,22 +101,17 @@ export const TopUp: React.FC = () => {
 
   React.useEffect(() => {
     handleSumupCardResp.current = (type: SumUpResponseType, body: object) => {
-      console.log("handle sumup resp called");
       if (state.stage !== "sumup") {
         return;
       }
-      console.log("Type", type);
-      console.log("Body", body);
       if (type === "invalid" && "message" in body && typeof body.message === "string") {
         toast.error(body.message);
       }
 
       if (type === "error" || type === "success") {
-        console.log("updating checkout");
-        checkCheckout({ checkCheckoutPayload: { checkout_id: state.checkoutId } })
+        checkCheckout({ checkCheckoutPayload: { order_uuid: state.orderUUID } })
           .unwrap()
           .then((resp) => {
-            console.log("update checkout returned with resp", resp);
             if (sumupCard.current) {
               sumupCard.current.unmount();
               sumupCard.current = undefined;
@@ -141,7 +140,6 @@ export const TopUp: React.FC = () => {
     if (state.stage !== "sumup") {
       return;
     }
-    console.log("checkoutId", state.checkoutId);
     const config = {
       id: "sumup-card",
       checkoutId: state.checkoutId,
@@ -150,7 +148,6 @@ export const TopUp: React.FC = () => {
       locale: i18n.language,
     };
     if (sumupCard.current) {
-      console.log("updating sumup card with config", config);
       sumupCard.current.update(config);
     } else {
       try {
@@ -181,8 +178,12 @@ export const TopUp: React.FC = () => {
     createCheckout({ createCheckoutPayload: values })
       .unwrap()
       .then((checkout) => {
-        console.log("created checkout with reference", checkout);
-        dispatch({ type: "created-checkout", checkoutId: checkout.checkout_id, topupAmount: values.amount });
+        dispatch({
+          type: "created-checkout",
+          checkoutId: checkout.checkout_id,
+          topupAmount: values.amount,
+          orderUUID: checkout.order_uuid,
+        });
         setSubmitting(false);
       })
       .catch((error) => {
